norns_defconfig: revert to separate rpi3/rpi4 configs
the nornsv7l_defconfig runs on both rpi3/rpi4 hardware but the
rpi3 incurs a ~3% cpu penalty due to the highmem and lpae options
which allow a 32-bit kernel to access >4GB ram on rpi4 hardware.
the norns_defconfig is not once again tuned for performance on rpi3
with the same memory addressing scheme as the stock rpi3 kernel
and previous 4.19.y kernel.